+++ web/src/components/item/ItemFields.tsx
@@ -5,9 +5,7 @@ import { FaAmazon } from 'react-icons/fa6';
 import { match } from 'ts-pattern';
 
 import { useItemFields } from '@/api';
-
-import { Button } from '../../gui/Button';
-import { FieldOption, FieldSelect } from '../../gui/select/FormSelect';
+import { Button, FieldOption, FieldSelect } from '@/gui';
 
 const AMAZON_DOMAINS = [
     { label: 'Amazon.com', value: 'amazon.com', icon: () => <FaAmazon /> },
