+++ web/src/components/item/all_elements.tsx
@@ -2,7 +2,7 @@ import { FC } from 'react';
 
 import { useFilteredItems } from '@/api';
 
-import { ItemPreview } from './item/ItemPreview';
+import { ItemPreview } from './ItemPreview';
 
 export const AllItems: FC<{ variant: 'full' | 'large' }> = ({
     variant,
