+++ web/src/routes/settings/_layout/sessions.tsx
@@ -1,6 +1,6 @@
 import { createFileRoute } from '@tanstack/react-router';
 
-import { ActiveSessionsTable } from '@/components/ActiveSessionsTable';
+import { ActiveSessionsTable } from '@/components/user/sessions/ActiveSessionsTable';
 
 export const Route = createFileRoute('/settings/_layout/sessions')({
     component: RouteComponent,
