+++ web/src/styles/index.css
@@ -2,6 +2,9 @@
 @tailwind components;
 @tailwind utilities;
 
+@import './animations.css';
+@import './scrollbar.css';
+
 @layer base {
     :root {
         --background: 0 0% 100%;
@@ -29,28 +32,6 @@
 
 #root {
     @apply min-h-screen flex flex-col;
-}
-
-/* Add this to your CSS file */
-/* For Webkit browsers (Chrome, Safari, etc.) */
-::-webkit-scrollbar {
-    width: 10px;
-    height: 10px;
-}
-
-::-webkit-scrollbar-track {
-    background: hsl(var(--muted) / 0.5);
-}
-
-::-webkit-scrollbar-thumb {
-    background: hsl(var(--muted-foreground) / 0.5);
-    border-radius: 100px;
-    border: 3px solid transparent; /* Add space around the thumb */
-    background-clip: padding-box;
-}
-
-::-webkit-scrollbar-thumb:hover {
-    background: hsl(var(--muted-foreground) / 0.7);
 }
 
 .btn {
@@ -165,49 +146,6 @@ a {
     font-weight: 500;
 }
 
-@keyframes slideUpAndFade {
-    0% {
-        opacity: 0;
-        transform: translateY(2px);
-    }
-    100% {
-        opacity: 1;
-        transform: translateY(0);
-    }
-}
-
-@keyframes slideRightAndFade {
-    0% {
-        opacity: 0;
-        transform: translateX(-2px);
-    }
-    100% {
-        opacity: 1;
-        transform: translateX(0);
-    }
-}
-
-@keyframes slideDownAndFade {
-    0% {
-        opacity: 0;
-        transform: translateY(-2px);
-    }
-    100% {
-        opacity: 1;
-        transform: translateY(0);
-    }
-}
-
-@keyframes slideLeftAndFade {
-    0% {
-        opacity: 0;
-        transform: translateX(2px);
-    }
-    100% {
-        opacity: 1;
-        transform: translateX(0);
-    }
-}
 
 .DropdownMenuContent,
 .DropdownMenuSubContent {
@@ -348,18 +286,6 @@ a {
     font-size: 15px;
     line-height: 1;
     @apply border rounded-md focus:outline focus:outline-2 outline-offset-2 outline-blue-500;
-}
-
-/* Define the keyframes for the loading animation */
-@keyframes loading {
-    0%,
-    80%,
-    100% {
-        opacity: 0;
-    }
-    40% {
-        opacity: 1;
-    }
 }
 
 /* Apply the loading animation to child spans */
